Extract more to projectsPicker
